feat(frontend): implement light mode
Adds light mode support into GUI, and button to switch between light and
dark mode in the navbar. User's preference is stored in the browser's
local storage for presistence. Browser's default setting is respected as
a initial value.

This commit also introduces changes to coloring of several elements
to maintain contrast in both light and dark mode.